Separate shortcut handling code into a ShortcutProvider widget
