+++ lib/widgets/editors/piano_roll/piano_roll_events.dart
@@ -18,7 +18,7 @@
 */
 
 import 'package:anthem/helpers/id.dart';
-import 'package:anthem/widgets/main_window/main_window_controller.dart';
+import 'package:anthem/widgets/basic/shortcuts/shortcut_provider.dart';
 import 'package:flutter/widgets.dart';
 
 abstract class PianoRollEvent {
